MA-5969 fix when we do factory reset,the colour of robot is blue.
The root cause is minui in recovery changes a lot on 5.0 which
remove PixelFlinger code and hard code to only support RGBX8888 format.
But 6SX and 6SL platform only support BGRX8888 and RGB565 format.

The solution is to provide RGB565 format because all imx6 serial platform
can support RGB565 format.

Signed-off-by: Liu Xiaowen <xiaowen.liu@freescale.com>